Move interfaces to Sdo.Interfaces namespace
Refactored IAuthenticationService and IPlatformDetector into a new Sdo.Interfaces namespace, removing them from Sdo.Services. Updated all references throughout the codebase to use the new namespace, improving separation of interfaces from their implementations and enhancing code organization.
